size_t and type consistency in crc32 algo, passes test. next commit will change name to not pollute namespace, which will break things

+++ src/checksum.h
@@ -10,7 +10,7 @@ extern "C" {
 
 
 uint16_t get_crc16(const unsigned char * arr, size_t size);
-uint32_t get_crc32(const unsigned char * message, uint32_t len);
+uint32_t get_crc32(const unsigned char * message, size_t len);
 
 #ifdef __cplusplus
 }
